+++ src/components/navigation/nav-menu/components/nav-icon-item/index.ts
@@ -1,2 +1,2 @@
 export type { NavigationLogoProps } from './nav-icon-item.types';
-export { NavigationLogo } from './nav-icon-item';
+export { NavigationLogo } from './navigation-logo';
