+++ src/components/navigation/nav-menu/components/nav-icon-item/navigation-logo.tsx
